refactor: split cli execution helpers
