+++ src/cli.rs
@@ -1,16 +1,15 @@
 use std::{
-    env,
     ffi::OsString,
     io::{self, Write},
     path::PathBuf,
-    thread,
-    time::Duration,
 };
 
 use chrono::{Datelike, NaiveDate};
 use serde::Serialize;
 
-use crate::app::{App, SetupCheck, SetupCheckLevel, SetupDiagnostics};
+#[cfg(test)]
+use crate::app::App;
+use crate::app::{SetupCheck, SetupCheckLevel, SetupDiagnostics};
 use crate::blocker::{BlockingPreviewAction, EditSiteResult, InvalidSiteInput, SiteBlocker};
 use crate::config::{
     AppConfig, BlocklistProfileConfig, BreakTemplateConfig, CustomProfileConfig, DailyGoalConfig,
@@ -29,8 +28,15 @@ use crate::timer::{
 };
 
 mod args;
+mod execute;
 mod output;
 use args::*;
+use execute::execute_cli_command;
+#[cfg(test)]
+use execute::{
+    apply_blocklist_profile_command, apply_site_add_command, apply_site_delete_command,
+    apply_site_edit_command,
+};
 use output::*;
 
 const USAGE_TEXT: &str = r#"Usage:
@@ -1087,989 +1093,7 @@ fn finalize_cli_action(
 }
 
 pub fn execute_command(cli_command: CliCommand) -> Result<(), String> {
-    match cli_command.kind {
-        CommandKind::Pause => execute_pause_command(cli_command.output),
-        CommandKind::Resume => execute_resume_command(cli_command.output),
-        CommandKind::Stop => execute_stop_command(cli_command.output),
-        CommandKind::Next => execute_next_command(cli_command.output),
-        CommandKind::Task { label } => execute_task_command(label, cli_command.output),
-        CommandKind::TaskGoal { label, goal } => {
-            execute_task_goal_command(label, goal, cli_command.output)
-        }
-        CommandKind::Profile { profile } => execute_profile_command(profile, cli_command.output),
-        CommandKind::Goal { goal } => execute_goal_command(goal, cli_command.output),
-        CommandKind::GoalWeekly { goal } => execute_weekly_goal_command(goal, cli_command.output),
-        CommandKind::GoalMonthly { goal } => execute_monthly_goal_command(goal, cli_command.output),
-        CommandKind::GoalCarry { enabled } => {
-            execute_goal_carry_command(enabled, cli_command.output)
-        }
-        CommandKind::GoalCarryWeekly { enabled } => {
-            execute_weekly_goal_carry_command(enabled, cli_command.output)
-        }
-        CommandKind::GoalCarryMonthly { enabled } => {
-            execute_monthly_goal_carry_command(enabled, cli_command.output)
-        }
-        CommandKind::Strict { enabled } => execute_strict_command(enabled, cli_command.output),
-        CommandKind::Schedule { schedule } => {
-            execute_schedule_command(schedule, cli_command.output)
-        }
-        CommandKind::Diagnostics => execute_diagnostics_command(cli_command.output),
-        CommandKind::BlockingPreview => execute_blocking_preview_command(cli_command.output),
-        CommandKind::Status {
-            watch_interval_secs,
-        } => execute_status_command(cli_command.output, watch_interval_secs),
-        CommandKind::Export { dir } => execute_export_command(dir, cli_command.output),
-        CommandKind::BlocklistProfile { command } => {
-            execute_blocklist_profile_command(command, cli_command.output)
-        }
-        CommandKind::BlocklistSites { target, command } => {
-            execute_blocklist_sites_command(target, command, cli_command.output)
-        }
-    }
-}
-
-fn execute_pause_command(output: OutputMode) -> Result<(), String> {
-    let mut app = App::new();
-    app.pause_for_cli()?;
-    emit_timer_command_output("pause", &app, output)
-}
-
-fn execute_resume_command(output: OutputMode) -> Result<(), String> {
-    let mut app = App::new();
-    app.resume_for_cli()?;
-    emit_timer_command_output("resume", &app, output)
-}
-
-fn execute_stop_command(output: OutputMode) -> Result<(), String> {
-    let mut app = App::new();
-    app.stop_for_cli()?;
-    emit_timer_command_output("stop", &app, output)
-}
-
-fn execute_next_command(output: OutputMode) -> Result<(), String> {
-    let mut app = App::new();
-    app.next_phase_for_cli()?;
-    emit_timer_command_output("next", &app, output)
-}
-
-fn execute_task_command(label: String, output: OutputMode) -> Result<(), String> {
-    let mut app = App::new();
-    let created = app.select_task_label_for_cli(&label)?;
-    let selected_task_label = app
-        .selected_task_label_for_cli()
-        .ok_or_else(|| "Task selection failed: no selected task label after update.".to_string())?;
-    let payload = TaskCommandOutput {
-        action: "task",
-        created,
-        selected_task_label,
-        timer: build_timer_state_output(&app),
-    };
-
-    match output {
-        OutputMode::Text => {
-            if payload.created {
-                println!(
-                    "Task label added and selected: {}",
-                    payload.selected_task_label
-                );
-            } else {
-                println!("Task label selected: {}", payload.selected_task_label);
-            }
-            print_timer_state_output(&payload.timer);
-        }
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_task_goal_command(
-    label: Option<String>,
-    goal: Option<DailyGoalConfig>,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut stats = FocusStats::load()?;
-    let selected_task_label = stats.task_planner_state().1;
-    let requested_label = label.or(selected_task_label).ok_or_else(|| {
-        "No task label selected. Use `--task LABEL` first or pass `--task-goal LABEL`.".to_string()
-    })?;
-
-    let mut updated = false;
-    let task_label = if let Some(goal) = goal {
-        let target = DailyGoalSnapshot {
-            minutes: goal.minutes,
-            pomodoros: goal.pomodoros,
-        };
-        let canonical = stats.set_task_goal_target(&requested_label, target)?;
-        stats
-            .save()
-            .map_err(|error| format!("Failed to save task goals: {error}"))?;
-        updated = true;
-        canonical
-    } else {
-        requested_label
-    };
-
-    let TaskGoalOutput {
-        task_label,
-        configured,
-        minutes_target,
-        pomodoros_target,
-        focused_minutes,
-        pomodoros_completed,
-        met,
-    } = build_task_goal_output(&stats, &task_label);
-    let payload = TaskGoalCommandOutput {
-        updated,
-        task_label,
-        configured,
-        minutes_target,
-        pomodoros_target,
-        focused_minutes,
-        pomodoros_completed,
-        met,
-    };
-
-    match output {
-        OutputMode::Text => print_task_goal_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_blocklist_profile_command(
-    command: BlocklistProfileCommandKind,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let payload = apply_blocklist_profile_command(&mut config, command)?;
-    if payload.updated {
-        config
-            .save()
-            .map_err(|error| format!("Failed to save blocklist profile settings: {error}"))?;
-    }
-
-    match output {
-        OutputMode::Text => print_blocklist_profile_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_blocklist_sites_command(
-    target: SiteListTarget,
-    command: BlocklistSiteCommandKind,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    match command {
-        BlocklistSiteCommandKind::List => {
-            let payload = build_site_list_command_output(&config, target, "site-list");
-            match output {
-                OutputMode::Text => print_site_list_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-        BlocklistSiteCommandKind::Add { input } => {
-            let payload = apply_site_add_command(&mut config, target, &input)?;
-            if payload.updated {
-                config
-                    .save()
-                    .map_err(|error| format!("Failed to save site changes: {error}"))?;
-            }
-            match output {
-                OutputMode::Text => print_site_add_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-        BlocklistSiteCommandKind::Edit { value } => {
-            let payload = apply_site_edit_command(&mut config, target, &value)?;
-            if payload.updated {
-                config
-                    .save()
-                    .map_err(|error| format!("Failed to save site changes: {error}"))?;
-            }
-            match output {
-                OutputMode::Text => print_site_edit_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-        BlocklistSiteCommandKind::Delete { site } => {
-            let payload = apply_site_delete_command(&mut config, target, &site)?;
-            if payload.updated {
-                config
-                    .save()
-                    .map_err(|error| format!("Failed to save site changes: {error}"))?;
-            }
-            match output {
-                OutputMode::Text => print_site_delete_command_output(&payload),
-                OutputMode::Json => print_json(&payload)?,
-            }
-        }
-    }
-    Ok(())
-}
-
-fn apply_blocklist_profile_command(
-    config: &mut AppConfig,
-    command: BlocklistProfileCommandKind,
-) -> Result<BlocklistProfileCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-
-    let (action, updated) = match command {
-        BlocklistProfileCommandKind::Select { profile } => {
-            handle_select_blocklist_profile(config, profile)?
-        }
-        BlocklistProfileCommandKind::Create { name } => {
-            handle_create_blocklist_profile(config, name)?
-        }
-        BlocklistProfileCommandKind::Rename { name } => {
-            handle_rename_blocklist_profile(config, name)?
-        }
-        BlocklistProfileCommandKind::Delete => handle_delete_blocklist_profile(config)?,
-    };
-
-    if updated {
-        sync_legacy_blocked_sites(config);
-    }
-    Ok(build_blocklist_profile_command_output(
-        config, action, updated,
-    ))
-}
-
-fn handle_select_blocklist_profile(
-    config: &mut AppConfig,
-    profile: Option<String>,
-) -> Result<(&'static str, bool), String> {
-    let mut updated = false;
-    if let Some(profile) = profile {
-        let index = blocklist_profile_index_by_name(&config.blocklist_profiles, &profile)
-            .ok_or_else(|| format!("Unknown blocklist profile `{profile}`."))?;
-        let selected = config.blocklist_profiles[index].name.clone();
-        if !config
-            .selected_blocklist_profile
-            .eq_ignore_ascii_case(&selected)
-        {
-            config.selected_blocklist_profile = selected;
-            updated = true;
-        }
-    }
-    Ok(("blocklist-profile", updated))
-}
-
-fn handle_create_blocklist_profile(
-    config: &mut AppConfig,
-    name: String,
-) -> Result<(&'static str, bool), String> {
-    let name = name.trim().to_string();
-    if name.is_empty() {
-        return Err("Profile name cannot be empty.".to_string());
-    }
-    if blocklist_profile_index_by_name(&config.blocklist_profiles, &name).is_some() {
-        return Err(format!("Profile `{name}` already exists."));
-    }
-    config.blocklist_profiles.push(BlocklistProfileConfig {
-        name: name.clone(),
-        sites: Vec::new(),
-        allowlist_sites: Vec::new(),
-    });
-    config.selected_blocklist_profile = name;
-    Ok(("blocklist-profile-create", true))
-}
-
-fn handle_rename_blocklist_profile(
-    config: &mut AppConfig,
-    name: String,
-) -> Result<(&'static str, bool), String> {
-    let index = selected_blocklist_profile_index(config);
-    let current_name = config.blocklist_profiles[index].name.clone();
-    let name = name.trim().to_string();
-    if name.is_empty() {
-        return Err("Profile name cannot be empty.".to_string());
-    }
-    if current_name == name {
-        return Ok(("blocklist-profile-rename", false));
-    }
-
-    let duplicate =
-        config
-            .blocklist_profiles
-            .iter()
-            .enumerate()
-            .any(|(candidate_index, profile)| {
-                candidate_index != index && profile.name.eq_ignore_ascii_case(&name)
-            });
-    if duplicate {
-        return Err(format!("Profile `{name}` already exists."));
-    }
-
-    config.blocklist_profiles[index].name = name.clone();
-    config.selected_blocklist_profile = name;
-    Ok(("blocklist-profile-rename", true))
-}
-
-fn handle_delete_blocklist_profile(config: &mut AppConfig) -> Result<(&'static str, bool), String> {
-    if config.blocklist_profiles.len() <= 1 {
-        return Err("At least one blocklist profile is required.".to_string());
-    }
-    let index = selected_blocklist_profile_index(config);
-    config.blocklist_profiles.remove(index);
-    let next_index = index.min(config.blocklist_profiles.len().saturating_sub(1));
-    config.selected_blocklist_profile = config.blocklist_profiles[next_index].name.clone();
-    Ok(("blocklist-profile-delete", true))
-}
-
-fn apply_site_add_command(
-    config: &mut AppConfig,
-    target: SiteListTarget,
-    input: &str,
-) -> Result<SiteAddCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    let mut working = SiteBlocker::new();
-    let existing_sites = active_profile_sites(config, index, target).to_vec();
-    for site in existing_sites {
-        working.add_site(site);
-    }
-    let result = working.add_sites_from_input(input);
-    let updated = !result.added.is_empty();
-    *active_profile_sites_mut(config, index, target) = working.sites.clone();
-    if updated {
-        sync_legacy_blocked_sites(config);
-    }
-
-    let active_profile = &config.blocklist_profiles[index];
-    Ok(SiteAddCommandOutput {
-        action: "site-add",
-        updated,
-        profile: active_profile.name.clone(),
-        target,
-        added: result.added,
-        duplicates: result.duplicates,
-        invalid: invalid_site_entries_output(&result.invalid),
-        sites: active_profile_sites(config, index, target).to_vec(),
-        effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile).len(),
-    })
-}
-
-fn apply_site_edit_command(
-    config: &mut AppConfig,
-    target: SiteListTarget,
-    value: &SiteEditValue,
-) -> Result<SiteEditCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    let previous = value.previous.trim();
-
-    let mut working = SiteBlocker::new();
-    for site in active_profile_sites(config, index, target) {
-        working.add_site(site.clone());
-    }
-    let edit_index = working
-        .sites
-        .iter()
-        .position(|site| site.eq_ignore_ascii_case(previous))
-        .ok_or_else(|| {
-            format!(
-                "Site `{}` was not found in {}.",
-                value.previous,
-                target.id()
-            )
-        })?;
-    let result = working.edit_site_from_input(edit_index, &value.next);
-    match result {
-        EditSiteResult::Updated { old, new } => {
-            *active_profile_sites_mut(config, index, target) = working.sites.clone();
-            sync_legacy_blocked_sites(config);
-            let active_profile = &config.blocklist_profiles[index];
-            Ok(SiteEditCommandOutput {
-                action: "site-edit",
-                updated: true,
-                profile: active_profile.name.clone(),
-                target,
-                previous: old,
-                current: new,
-                sites: active_profile_sites(config, index, target).to_vec(),
-                effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile)
-                    .len(),
-            })
-        }
-        EditSiteResult::Unchanged { hostname } => {
-            let active_profile = &config.blocklist_profiles[index];
-            Ok(SiteEditCommandOutput {
-                action: "site-edit",
-                updated: false,
-                profile: active_profile.name.clone(),
-                target,
-                previous: hostname.clone(),
-                current: hostname,
-                sites: active_profile_sites(config, index, target).to_vec(),
-                effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile)
-                    .len(),
-            })
-        }
-        EditSiteResult::Duplicate { hostname } => {
-            Err(format!("`{hostname}` already exists in {}.", target.id()))
-        }
-        EditSiteResult::Invalid(invalid) => Err(format!(
-            "Invalid hostname `{}` ({})",
-            display_input_value(&invalid.input),
-            invalid.reason.message()
-        )),
-        EditSiteResult::MissingSelection => Err("No site selected to edit.".to_string()),
-    }
-}
-
-fn apply_site_delete_command(
-    config: &mut AppConfig,
-    target: SiteListTarget,
-    site: &str,
-) -> Result<SiteDeleteCommandOutput, String> {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    let site = site.trim();
-
-    let mut working = SiteBlocker::new();
-    for current in active_profile_sites(config, index, target) {
-        working.add_site(current.clone());
-    }
-    let delete_index = working
-        .sites
-        .iter()
-        .position(|value| value.eq_ignore_ascii_case(site))
-        .ok_or_else(|| format!("Site `{site}` was not found in {}.", target.id()))?;
-    let removed = working
-        .remove_site(delete_index)
-        .ok_or_else(|| format!("Site `{site}` was not found in {}.", target.id()))?;
-    *active_profile_sites_mut(config, index, target) = working.sites.clone();
-    sync_legacy_blocked_sites(config);
-
-    let active_profile = &config.blocklist_profiles[index];
-    Ok(SiteDeleteCommandOutput {
-        action: "site-delete",
-        updated: true,
-        profile: active_profile.name.clone(),
-        target,
-        removed,
-        sites: active_profile_sites(config, index, target).to_vec(),
-        effective_blocked_sites_count: effective_blocked_sites_for_profile(active_profile).len(),
-    })
-}
-
-fn ensure_blocklist_profiles(config: &mut AppConfig) {
-    if config.blocklist_profiles.is_empty() {
-        config
-            .blocklist_profiles
-            .push(BlocklistProfileConfig::default());
-        config.selected_blocklist_profile = config.blocklist_profiles[0].name.clone();
-    }
-}
-
-fn blocklist_profile_index_by_name(
-    profiles: &[BlocklistProfileConfig],
-    name: &str,
-) -> Option<usize> {
-    profiles
-        .iter()
-        .position(|profile| profile.name.eq_ignore_ascii_case(name.trim()))
-}
-
-fn selected_blocklist_profile_index(config: &AppConfig) -> usize {
-    blocklist_profile_index_by_name(
-        &config.blocklist_profiles,
-        &config.selected_blocklist_profile,
-    )
-    .unwrap_or(0)
-}
-
-fn active_profile_sites(
-    config: &AppConfig,
-    profile_index: usize,
-    target: SiteListTarget,
-) -> &[String] {
-    match target {
-        SiteListTarget::Blocklist => &config.blocklist_profiles[profile_index].sites,
-        SiteListTarget::Allowlist => &config.blocklist_profiles[profile_index].allowlist_sites,
-    }
-}
-
-fn active_profile_sites_mut(
-    config: &mut AppConfig,
-    profile_index: usize,
-    target: SiteListTarget,
-) -> &mut Vec<String> {
-    match target {
-        SiteListTarget::Blocklist => &mut config.blocklist_profiles[profile_index].sites,
-        SiteListTarget::Allowlist => &mut config.blocklist_profiles[profile_index].allowlist_sites,
-    }
-}
-
-fn sync_legacy_blocked_sites(config: &mut AppConfig) {
-    ensure_blocklist_profiles(config);
-    let index = selected_blocklist_profile_index(config);
-    config.selected_blocklist_profile = config.blocklist_profiles[index].name.clone();
-    config.blocked_sites = effective_blocked_sites_for_profile(&config.blocklist_profiles[index]);
-}
-
-fn build_blocklist_profile_command_output(
-    config: &AppConfig,
-    action: &'static str,
-    updated: bool,
-) -> BlocklistProfileCommandOutput {
-    let selected_name = config
-        .blocklist_profiles
-        .get(selected_blocklist_profile_index(config))
-        .map(|profile| profile.name.clone())
-        .unwrap_or_else(|| config.selected_blocklist_profile.clone());
-    let profiles = config
-        .blocklist_profiles
-        .iter()
-        .map(|profile| BlocklistProfileSummaryOutput {
-            name: profile.name.clone(),
-            active: profile.name.eq_ignore_ascii_case(&selected_name),
-            blocklist_sites_count: profile.sites.len(),
-            allowlist_sites_count: profile.allowlist_sites.len(),
-            effective_blocked_sites_count: effective_blocked_sites_for_profile(profile).len(),
-        })
-        .collect();
-
-    BlocklistProfileCommandOutput {
-        action,
-        updated,
-        selected_blocklist_profile: selected_name,
-        profiles,
-    }
-}
-
-fn build_site_list_command_output(
-    config: &AppConfig,
-    target: SiteListTarget,
-    action: &'static str,
-) -> SiteListCommandOutput {
-    if config.blocklist_profiles.is_empty() {
-        let fallback = if config.selected_blocklist_profile.trim().is_empty() {
-            "Default".to_string()
-        } else {
-            config.selected_blocklist_profile.clone()
-        };
-        return SiteListCommandOutput {
-            action,
-            profile: fallback,
-            target,
-            sites: Vec::new(),
-            effective_blocked_sites_count: 0,
-        };
-    }
-    let index = selected_blocklist_profile_index(config);
-    let profile = &config.blocklist_profiles[index];
-    SiteListCommandOutput {
-        action,
-        profile: profile.name.clone(),
-        target,
-        sites: active_profile_sites(config, index, target).to_vec(),
-        effective_blocked_sites_count: effective_blocked_sites_for_profile(profile).len(),
-    }
-}
-
-fn invalid_site_entries_output(values: &[InvalidSiteInput]) -> Vec<InvalidSiteEntryOutput> {
-    values
-        .iter()
-        .map(|invalid| InvalidSiteEntryOutput {
-            input: invalid.input.clone(),
-            reason: invalid.reason.message().to_string(),
-        })
-        .collect()
-}
-
-fn execute_profile_command(profile: Option<ProfileId>, output: OutputMode) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(profile) = profile {
-        config.selected_profile = profile;
-        config.align_legacy_automation_with_selected_profile();
-        config
-            .save()
-            .map_err(|error| format!("Failed to save selected profile: {error}"))?;
-        updated = true;
-    }
-
-    let custom = config.effective_custom_profile();
-    let selected = profile_view(config.selected_profile, &custom);
-    let available = [ProfileId::Classic, ProfileId::DeepWork, ProfileId::Custom]
-        .into_iter()
-        .map(|candidate| profile_view(candidate, &custom))
-        .collect();
-    let selected_break_template = selected_break_template_view(&config);
-    let available_break_templates = available_break_template_views(&config);
-    let payload = ProfileOutput {
-        updated,
-        selected,
-        available,
-        selected_break_template,
-        available_break_templates,
-    };
-
-    match output {
-        OutputMode::Text => print_profile_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_goal_command(goal: Option<DailyGoalConfig>, output: OutputMode) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(goal) = goal {
-        config.daily_goal = goal;
-        config
-            .save()
-            .map_err(|error| format!("Failed to save daily goal: {error}"))?;
-        updated = true;
-    }
-
-    let payload = GoalCommandOutput {
-        updated,
-        configured: config.daily_goal.minutes > 0 || config.daily_goal.pomodoros > 0,
-        minutes_target: config.daily_goal.minutes,
-        pomodoros_target: config.daily_goal.pomodoros,
-    };
-
-    match output {
-        OutputMode::Text => print_goal_command_output("Daily", &payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_weekly_goal_command(
-    goal: Option<WeeklyGoalConfig>,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(goal) = goal {
-        config.weekly_goal = goal;
-        config
-            .save()
-            .map_err(|error| format!("Failed to save weekly goal: {error}"))?;
-        updated = true;
-    }
-
-    let payload = GoalCommandOutput {
-        updated,
-        configured: config.weekly_goal.minutes > 0 || config.weekly_goal.pomodoros > 0,
-        minutes_target: config.weekly_goal.minutes,
-        pomodoros_target: config.weekly_goal.pomodoros,
-    };
-
-    match output {
-        OutputMode::Text => print_goal_command_output("Weekly", &payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_monthly_goal_command(
-    goal: Option<MonthlyGoalConfig>,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(goal) = goal {
-        config.monthly_goal = goal;
-        config
-            .save()
-            .map_err(|error| format!("Failed to save monthly goal: {error}"))?;
-        updated = true;
-    }
-
-    let payload = GoalCommandOutput {
-        updated,
-        configured: config.monthly_goal.minutes > 0 || config.monthly_goal.pomodoros > 0,
-        minutes_target: config.monthly_goal.minutes,
-        pomodoros_target: config.monthly_goal.pomodoros,
-    };
-
-    match output {
-        OutputMode::Text => print_goal_command_output("Monthly", &payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_goal_carry_command(enabled: Option<bool>, output: OutputMode) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(enabled) = enabled {
-        config.goal_carry_over.daily = enabled;
-        config
-            .save()
-            .map_err(|error| format!("Failed to save daily goal carry-over: {error}"))?;
-        updated = true;
-    }
-
-    let payload = GoalCarryCommandOutput {
-        updated,
-        carry_over: config.goal_carry_over.daily,
-    };
-    match output {
-        OutputMode::Text => print_goal_carry_command_output("Daily", &payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_weekly_goal_carry_command(
-    enabled: Option<bool>,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(enabled) = enabled {
-        config.goal_carry_over.weekly = enabled;
-        config
-            .save()
-            .map_err(|error| format!("Failed to save weekly goal carry-over: {error}"))?;
-        updated = true;
-    }
-
-    let payload = GoalCarryCommandOutput {
-        updated,
-        carry_over: config.goal_carry_over.weekly,
-    };
-    match output {
-        OutputMode::Text => print_goal_carry_command_output("Weekly", &payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_monthly_goal_carry_command(
-    enabled: Option<bool>,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(enabled) = enabled {
-        config.goal_carry_over.monthly = enabled;
-        config
-            .save()
-            .map_err(|error| format!("Failed to save monthly goal carry-over: {error}"))?;
-        updated = true;
-    }
-
-    let payload = GoalCarryCommandOutput {
-        updated,
-        carry_over: config.goal_carry_over.monthly,
-    };
-    match output {
-        OutputMode::Text => print_goal_carry_command_output("Monthly", &payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_strict_command(enabled: Option<bool>, output: OutputMode) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(enabled) = enabled {
-        let mut automation = config.profile_automation_for(config.selected_profile);
-        automation.strict_mode = enabled;
-        config.set_profile_automation_for(config.selected_profile, automation);
-        config
-            .save()
-            .map_err(|error| format!("Failed to save strict mode: {error}"))?;
-        updated = true;
-    }
-
-    let payload = StrictCommandOutput {
-        updated,
-        strict_mode: config
-            .profile_automation_for(config.selected_profile)
-            .strict_mode,
-    };
-
-    match output {
-        OutputMode::Text => print_strict_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_schedule_command(
-    schedule: Option<RecurringScheduleConfig>,
-    output: OutputMode,
-) -> Result<(), String> {
-    let mut config = AppConfig::load().normalized();
-    let mut updated = false;
-    if let Some(schedule) = schedule {
-        let mut automation = config.profile_automation_for(config.selected_profile);
-        automation.recurring_schedule = schedule.normalized();
-        config.set_profile_automation_for(config.selected_profile, automation);
-        config
-            .save()
-            .map_err(|error| format!("Failed to save recurring schedule: {error}"))?;
-        updated = true;
-    }
-    let selected_automation = config.profile_automation_for(config.selected_profile);
-    let inspection = build_schedule_inspection_output(&selected_automation.recurring_schedule);
-    let payload = ScheduleCommandOutput {
-        updated,
-        schedule: selected_automation.recurring_schedule,
-        inspection,
-    };
-
-    match output {
-        OutputMode::Text => print_schedule_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_diagnostics_command(output: OutputMode) -> Result<(), String> {
-    let app = App::new();
-    let payload = build_diagnostics_command_output(&app.setup_diagnostics);
-
-    match output {
-        OutputMode::Text => print_diagnostics_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_blocking_preview_command(output: OutputMode) -> Result<(), String> {
-    let app = App::new();
-    let preview = app.blocking_preview_for_cli()?;
-    let payload = build_blocking_preview_command_output(&preview);
-
-    match output {
-        OutputMode::Text => print_blocking_preview_command_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn execute_status_command(
-    output: OutputMode,
-    watch_interval_secs: Option<u64>,
-) -> Result<(), String> {
-    if let Some(interval_secs) = watch_interval_secs {
-        return execute_status_watch_command(output, interval_secs);
-    }
-
-    let payload = load_status_output()?;
-    emit_status_output(&payload, output, false)
-}
-
-fn execute_status_watch_command(output: OutputMode, interval_secs: u64) -> Result<(), String> {
-    if interval_secs == 0 {
-        return Err("`--watch` interval must be greater than 0 seconds.".to_string());
-    }
-
-    loop {
-        let payload = load_status_output()?;
-        emit_status_output(&payload, output, true)?;
-        flush_stdout()?;
-        thread::sleep(Duration::from_secs(interval_secs));
-    }
-}
-
-fn load_status_output() -> Result<StatusOutput, String> {
-    let config = AppConfig::load().normalized();
-    let stats = FocusStats::load().map_err(|error| format!("Failed to load stats: {error}"))?;
-    Ok(build_status_output(&config, &stats))
-}
-
-fn emit_status_output(
-    payload: &StatusOutput,
-    output: OutputMode,
-    watch_mode: bool,
-) -> Result<(), String> {
-    match output {
-        OutputMode::Text => {
-            print_status_output(payload);
-            if watch_mode {
-                println!();
-            }
-        }
-        OutputMode::Json => {
-            if watch_mode {
-                print_json_compact(payload)?;
-            } else {
-                print_json(payload)?;
-            }
-        }
-    }
-    Ok(())
-}
-
-fn execute_export_command(dir: Option<PathBuf>, output: OutputMode) -> Result<(), String> {
-    let stats = FocusStats::load().map_err(|error| format!("Failed to load stats: {error}"))?;
-    let target_dir = match dir {
-        Some(path) => path,
-        None => env::current_dir()
-            .map_err(|error| format!("Failed to determine current directory: {error}"))?,
-    };
-    let exported = stats
-        .export_to_dir(&target_dir)
-        .map_err(|error| format!("Export failed: {error}"))?;
-
-    let payload = ExportOutput {
-        export_dir: target_dir,
-        json_path: exported.json_path,
-        csv_path: exported.csv_path,
-    };
-    match output {
-        OutputMode::Text => print_export_output(&payload),
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn emit_timer_command_output(
-    action: &'static str,
-    app: &App,
-    output: OutputMode,
-) -> Result<(), String> {
-    let payload = TimerCommandOutput {
-        action,
-        timer: build_timer_state_output(app),
-    };
-
-    match output {
-        OutputMode::Text => {
-            println!("Timer action applied: {}", payload.action);
-            print_timer_state_output(&payload.timer);
-        }
-        OutputMode::Json => print_json(&payload)?,
-    }
-    Ok(())
-}
-
-fn build_timer_state_output(app: &App) -> TimerStateOutput {
-    let (phase, status, remaining_secs, pomodoros_completed) = app.timer_state_for_cli();
-    let profile = app.selected_profile_id();
-    let (focus_secs, short_break_secs, long_break_secs, long_break_interval) =
-        app.profile_values(profile);
-    let (selected_task_label, focus_intention, task_note) =
-        mirror_metadata_from_task_label(app.selected_task_label_for_cli());
-
-    TimerStateOutput {
-        phase: timer_phase_id(phase),
-        status: timer_status_id(status),
-        remaining_secs,
-        pomodoros_completed,
-        selected_profile: ProfileView {
-            id: profile_id(profile),
-            label: profile.label(),
-            focus_secs,
-            short_break_secs,
-            long_break_secs,
-            long_break_interval,
-        },
-        selected_task_label,
-        focus_intention,
-        task_note,
-    }
+    execute_cli_command(cli_command)
 }
 
 fn build_status_output(config: &AppConfig, stats: &FocusStats) -> StatusOutput {
